Begin to style the webpages buttons and headlines
Fix some typos and remove unnecessary stuff on index.html. Add style to
the sub-button as well as to the headlines for each webpage and add a
div to structure the headline better and make it look better. For that
create a class to set rules for every headline and sub-button on the
webpage